image property fix
git-svn-id: http://kanon-framework.googlecode.com/svn/trunk@1584 05b5fde2-2f63-11df-a735-6ff41f51f3aa

+++ src/mvc-model/properties/imageFilenameProperty.php
@@ -110,9 +110,11 @@ class imageFilenameProperty extends stringProperty{
     }
     public function preUpdate(){
         // if ($this->_changed){
-        $path = $this->getPath().'.thumb/';
-        foreach (glob($path.'tm*_'.$this->getValue()) as $filename){
-            unlink($filename);
+        if ($this->hasChangedValue()){
+            $path = $this->getPath().'.thumb/';
+            foreach (glob($path.'tm*_'.$this->getValue()) as $filename){
+                unlink($filename);
+            }
         }
         // }
     }
